Logistic_Regression_model/requirements.txt
flask
torch
torchvision
facenet-pytorch
numpy
joblib
scikit-learn
Pillow